move things around to a more proper format
