+++ src/error.rs
@@ -1,0 +1,10 @@
+/// Simple Error struct that has a String value for whatever reason it errored.
+// TODO rewrite as enum
+#[derive(Debug, Clone)]
+pub struct Error(String);
+
+impl fmt::Display for Error {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "Error: {}", self.0)
+    }
+}
